+++ src/API/CLIPPY/ClippyLoader.service.ts
@@ -1,9 +1,10 @@
 import clippy, { type Agent } from 'clippyts';
-export default (agent: string) =>
-  new Promise<Agent>((resolve, reject) =>
+export default function (agent: string) {
+  return new Promise<Agent>((resolve, reject) =>
     clippy.load({
       name: agent,
       successCb: (agent: Agent) => {
+        console.log('clippy loaded', { agent })
         resolve(agent);
       },
       failCb: (error) => {
@@ -13,3 +14,4 @@ export default (agent: string) =>
     )
 
   )
+}