+++ src/Component/App/Stateless/Style/Theme/LIGHT.theme.ts
@@ -92,6 +92,13 @@ const LIGHT = createTheme({
         },
       }
     },
+    MuiCheckbox: {
+      styleOverrides: {
+        sizeSmall: {
+          '>.MuiSvgIcon-root': { fontSize: '1.2rem' }
+        }
+      }
+    },
     MuiChip: {
       styleOverrides: {
         clickable: { '&:hover': { backgroundColor: grey[400] } },