+++ src/Effect/usePriceFieldAdornment.ts
@@ -1,13 +1,18 @@
 import { SxProps } from '@mui/material';
 
+export const AdornmentPosition = {
+  start: 'start',
+  end: 'end'
+} as const
+export type AdornmentPosition = typeof AdornmentPosition[keyof typeof AdornmentPosition]
 const styles = { userSelect: 'none' }
-const usePriceFieldAdornment = (currency: string): ['endAdornment' | 'startAdornment', 'start' | 'end', SxProps] => {
+const usePriceFieldAdornment = (variant: AdornmentPosition): ['endAdornment' | 'startAdornment', 'start' | 'end', SxProps] => {
   const adornmentType =
-    currency.toLowerCase() === 'eur'
+    variant == AdornmentPosition.end
       ? 'endAdornment'
       : 'startAdornment'
   const adornmentPosition =
-    currency.toLowerCase() === 'eur'
+    variant == AdornmentPosition.end
       ? 'end'
       : 'start'
   return [adornmentType, adornmentPosition, styles];