+++ src/Store/Epic/Clippy/index.ts
@@ -1,3 +1,3 @@
 import { combineEpics } from 'redux-observable';
-import save from './save.epic'
+import save from './saveSelection.epic'
 export default combineEpics(save)